% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download-repos.R
\name{download_all}
\alias{download_all}
\title{Download All GitHub Repositories For A User}
\usage{
download_all(gh_user, dir)
}
\arguments{
\item{gh_user}{Character string. A GitHub user name.}

\item{dir}{Character string. A local file path where the zipped repositories
will be downloaded to.}
}
\value{
Unzipped GitHub repositories in a specified local directory.
}
\description{
Download all of a GitHub user's repositories to your computer as zip files
and then optionally unzip them. Make sure you've got a GitHub account and
have generated a GitHub PAT token and stored it your .Renviron. See
package README for details.
}
\examples{
\dontrun{
download_all(gh_user = "matt-dray", dir = "~/Documents/repos")
}
}
