% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download-repos.R
\name{download_repo_zips}
\alias{download_repo_zips}
\title{Download Zipped Repos To Your Machine}
\usage{
download_repo_zips(repo_urls, dest_dir)
}
\arguments{
\item{repo_urls}{A tibble/data.frame as returned by
\code{\link{enframe_repo_urls}}, with one row per GitHub repo with
character-class columns for the repo_name and zip_url.}

\item{dest_dir}{A character string. The local directory you want to download
the zipped files to.}
}
\value{
Zipped GitHub repositories downloaded to a (possibly new) directory.
}
\description{
Download to a specified location the zip files from provided URL. The
directory is created if it doesn't already exist.
}
