% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download-repos.R
\name{enframe_repo_urls}
\alias{enframe_repo_urls}
\title{Create A Data Frame Of Repo Names And Zip File URLS For Each}
\usage{
enframe_repo_urls(repo_names, gh_user)
}
\arguments{
\item{repo_names}{A character vector of GitHub repository names, as returned
by \code{\link{extract_repo_names}}.}

\item{gh_user}{The name of a GitHub user.}
}
\value{
A tibble/data.frame object. One row per GitHub repo with character
vector columns for the repo_name and zip_url.
}
\description{
Prepare a data frame containing each repo name and its corresponding zip file
URL in the form https://github.com/username/reponame/archive/master.zip
}
