% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download-repos.R
\name{extract_repo_names}
\alias{extract_repo_names}
\title{Extract Names of GitHub Repos}
\usage{
extract_repo_names(repo_object)
}
\arguments{
\item{repo_object}{A gh_response object, as returned by \code{\link{get_repos}}.}
}
\value{
A character vector of GitHub repo names.
}
\description{
Extract all the 'name' elements from a gh_response object. These are the
names of all the GitHub repos.
}
