% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04-copy-repos.R
\name{ghd_copy}
\alias{ghd_copy}
\title{Clone Or Download All GitHub Repositories For A User}
\usage{
ghd_copy(gh_user, dest_dir, copy_type = c("download", "clone"))
}
\arguments{
\item{gh_user}{Character string. A GitHub user name.}

\item{dest_dir}{Character string. A local file path where the zipped repositories
will be downloaded to.}

\item{copy_type}{Character string. Specify whether to \code{"download"}
or \code{"clone"} the repos to your local machine.}
}
\value{
GitHub repositories either (a) downloaded or (b) cloned in the
specified local directory.
}
\description{
Clone all of a GitHub user's repositories to your computer, or download them
all as zip files and then optionally unzip them. Make sure you've got a
GitHub account and have generated a GitHub PAT token and stored it your
.Renviron. See package README for details.
}
\examples{
\dontrun{
ghd_get(
  gh_user = "matt-dray",
  dest_dir = "~/Documents/repos",
  copy_type = "clone"
)
}
}
