% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-download-repos.R
\name{ghd_download_zips}
\alias{ghd_download_zips}
\title{Download Zipped Repos To Your Machine}
\usage{
ghd_download_zips(repo_urls, dest_dir)
}
\arguments{
\item{repo_urls}{A data.frame as returned by
\code{\link{ghd_enframe_urls}}, with one row per GitHub repo with
character-class columns for the repo_names and zip_url.}

\item{dest_dir}{A character string. The local directory you want to download
the zipped files to. Must be a full path.}
}
\value{
Zipped GitHub repositories downloaded to a (possibly new) directory.
}
\description{
Download to a specified location the zip files from provided URL. The
directory is created if it doesn't already exist.
}
