% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-repo-names.R
\name{ghd_get_repos}
\alias{ghd_get_repos}
\title{Get GitHub Repo Information for a Named User}
\usage{
ghd_get_repos(gh_user)
}
\arguments{
\item{gh_user}{Character string. A GitHub username.}
}
\value{
A gh_response object.
}
\description{
Uses \code{\link[gh]{gh}} to access the GitHub API and get the details for
all of a named user's repos. To use this function you need to have created a
GitHub account and to have put a GitHub Personal Access Token (PAT) in your
.Renviron. See \href{https://happygitwithr.com/github-pat.html}{Happy Git and GitHub for the UseR}
for omre information.
}
