% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-download-repos.R
\name{ghd_unzip}
\alias{ghd_unzip}
\title{Unzip GitHub Repositories In A Named Directory}
\usage{
ghd_unzip(dir)
}
\arguments{
\item{dir}{A string. Path to a local directory containing zipped GitHub
directories. These may have been downloaded using
\code{\link{ghd_download_zips}}. Must be a full path.}
}
\value{
Unzipped GitHub repositories in a named directory.
}
\description{
Unzips the GitHub repositories that have been downloaded to a specified
local directory and optionally deletes the zipped versions. Optionally
rename all the unzipped folders to remove the "-master" suffix (e.g.
"demo-repo-master" becomes "demo-repo").
}
