% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghdump-package.R
\docType{package}
\name{ghdump-package}
\alias{ghdump}
\alias{ghdump-package}
\title{ghdump: Download All Repos For A GitHub User}
\description{
Clone all of a GitHub user's repositories, or download them as zip
    files to a specified location and optionally unzip them. Intended for
    archiving or setting up a new computer.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/matt-dray/ghdump/}
  \item Report bugs at \url{https://github.com/matt-dray/ghdump/issues}
}

}
\author{
\strong{Maintainer}: Matt Dray \email{mwdray@gmail.com}

}
\keyword{internal}
