% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download-repos.R
\name{unzip_repos}
\alias{unzip_repos}
\title{Unzip GitHub Repositories In A Named Directory}
\usage{
unzip_repos(dir, rm_zip = TRUE, rename_dir = TRUE)
}
\arguments{
\item{dir}{A string. Path to a local directory containing zipped GitHub
directories. These may have been downloaded using
\code{\link{download_repo_zips}}.}

\item{rm_zip}{Logical. Having unzipped the files, should the zipped versions
be deleted?}

\item{rename_dir}{Logical. Should the unzipped files be renamed to remove the
"-master" suffix?}
}
\value{
Unzipped GitHub repositories in a named directory.
}
\description{
Unzips the GitHub repositories that have been downloaded to a specified
local directory and optionally deletes the zipped versions. Optionally
rename all the unzipped folders to remove the "-master" suffix (e.g.
"demo-repo-master" becomes "demo-repo").
}
