% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-clone-repos.R
\name{ghd_clone_one}
\alias{ghd_clone_one}
\title{Clone One Repo From A GitHub User}
\usage{
ghd_clone_one(gh_user, repo, dest_dir)
}
\arguments{
\item{gh_user}{Character string. A GitHub user name.}

\item{repo}{Character string. A GitHub repo name for the named \code{gh_user}.}

\item{dest_dir}{Character string. A local file path where the zipped repositories
will be downloaded to.}
}
\value{
The named user's named repo cloned to the specified location.
}
\description{
Clone a repo for a specified user to a specified local repository.
}
